Remove hard coding of source into Controller.java, a little refactoring dealing with that
