--- proj9DouglasHanssenMacDonaldZhang/A.java
@@ -1,29 +1,0 @@
-/*
-Tia Zhang wrote this test file 
-*/
-
-//For project 9
-
-/*
-And she's paranoid // about whether it will work correctly
-So she's
-//trying to test thoroughly
-*/
-
-class A {
-	void main(String[] args){
-	while(true) {
-		int i=0;
-		System.out.println("Hello, world");
-		Thread.sleep(500);
-		i++;
-		i--;
-		if(i==true)
-		if(i==false)
-		if(i!=blargle)
-		String blah = " Hi
-			Bob";
-		String bob = "\0"
-		}
-	}
-}
